move shared types to their own crate to limit dependencies
