+++ Dockerfile
@@ -35,6 +35,7 @@ RUN curl https://sh.rustup.rs -sSf | sh -s -- -y --default-toolchain nightly
 ENV PATH=/home/user/.cargo/bin:/home/user/bin:$PATH
 
 COPY . /home/user
+WORKDIR /home/user/server
 RUN cargo build --release
 
 USER root
