+++ api/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "registration_types"
+version = "0.1.0"
+authors = ["fabrice <fabrice@desre.org>"]
+
+[dependencies]
+serde = "1.0"
+serde_derive = "1.0"
