+++ server/Cargo.toml
@@ -1,0 +1,31 @@
+[package]
+name = "registration_server"
+version = "0.1.0"
+authors = ["fabrice <fabrice@desre.org>"]
+
+[dependencies]
+clap = "2"
+email = "0.0"
+env_logger = "0.3"
+hyper-openssl = "0.2"
+iron = "0.5"
+iron-cors = { git = "https://github.com/fxbox/iron-cors.git" }
+lettre = "0.6"
+log = "0.3"
+params = "0.6"
+mount = "0.3"
+registration_types = { path = "../api" }
+r2d2 = "0.7"
+r2d2_sqlite = "0.2"
+rusqlite = "0.11"
+router = "0.5"
+rust-crypto = "0.2"
+serde = "1.0"
+serde_derive = "1.0"
+serde_json = "1.0"
+toml = "0.4"
+uuid = { version = "0.4", features = ["v4"] }
+
+[dev-dependencies]
+iron-test = "0.5"
+hyper = "0.10"
