+++ server/src/email_routes.rs
